imrpove mcp connect tutorial
