+++ src/locale/de.json
@@ -3309,6 +3309,43 @@
           "exportFeed": ""
         },
         "claude": {
+          "requirements": {
+            "title": "",
+            "description": "",
+            "checkFirstTitle": "",
+            "checkFirstDescription": "",
+            "skipIfVersions": "",
+            "nodeDownloadLabel": "",
+            "tabs": {
+              "mac": "",
+              "windows": ""
+            },
+            "mac": {
+              "title": "",
+              "steps": {
+                "openTerminal": "",
+                "runArchitectureCommand": "",
+                "chooseArchitecture": "",
+                "downloadInstaller": "",
+                "runInstaller": "",
+                "verifyAgain": ""
+              },
+              "appleSilicon": "",
+              "intel": ""
+            },
+            "windows": {
+              "title": "",
+              "steps": {
+                "openCommandPrompt": "",
+                "openSettings": "",
+                "openAbout": "",
+                "chooseInstaller": "",
+                "downloadInstaller": "",
+                "runInstaller": "",
+                "verifyAgain": ""
+              }
+            }
+          },
           "title": "",
           "instructions": "",
           "fullExampleTitle": "",
@@ -3317,6 +3354,7 @@
         "exportFeed": {
           "title": "",
           "description": "",
+          "curlTitle": "",
           "pythonTitle": "",
           "phpTitle": ""
         }
