+++ src/locale/en.json
@@ -1193,6 +1193,43 @@
           "exportFeed": "Export Feed"
         },
         "claude": {
+          "requirements": {
+            "title": "Before you set up Claude",
+            "description": "Claude needs Node.js on your computer before it can use the OneSila connection. Start with the quick check below.",
+            "checkFirstTitle": "First, check if you already have it",
+            "checkFirstDescription": "On a Mac, press Command + Space, type Terminal, and press Enter. On Windows, click Start, type Command Prompt, and press Enter. Then copy and paste these commands one at a time.",
+            "skipIfVersions": "If all three commands show version numbers, you can skip the install steps and continue with the Claude setup below.",
+            "nodeDownloadLabel": "Open the Node.js download page here:",
+            "tabs": {
+              "mac": "MacBook",
+              "windows": "Windows"
+            },
+            "mac": {
+              "title": "Install Node.js on a MacBook",
+              "steps": {
+                "openTerminal": "Open Terminal: press Command + Space on your keyboard, type Terminal, then press Enter.",
+                "runArchitectureCommand": "Run the command below to check your Mac type.",
+                "chooseArchitecture": "If it says arm64, choose Apple Silicon. If it says x86_64, choose Intel.",
+                "downloadInstaller": "Go to the Node.js download page, choose macOS, choose Installer, and select the architecture that matches your Mac.",
+                "runInstaller": "After the download finishes, open the installer file and follow the steps on screen.",
+                "verifyAgain": "When the installer is done, open Terminal again with Command + Space, type Terminal, press Enter, and run the three version commands to check that it worked."
+              },
+              "appleSilicon": "arm64 means your Mac uses Apple Silicon.",
+              "intel": "x86_64 means your Mac uses Intel."
+            },
+            "windows": {
+              "title": "Install Node.js on Windows",
+              "steps": {
+                "openCommandPrompt": "Open Command Prompt first: click the Start button at the bottom of the screen, type Command Prompt, then press Enter.",
+                "openSettings": "Open Settings from the Start menu.",
+                "openAbout": "Go to System, then About, and look for System type.",
+                "chooseInstaller": "If System type says ARM-based processor or ARM64, choose Windows ARM64 Installer. Otherwise, choose Windows x64 Installer.",
+                "downloadInstaller": "Go to the Node.js download page and download the correct Windows Installer (.msi).",
+                "runInstaller": "Open the downloaded .msi file and follow the steps on screen.",
+                "verifyAgain": "After installation, open Command Prompt: click Start, type Command Prompt, press Enter, and run the three version commands to check that it worked."
+              }
+            }
+          },
           "title": "Claude Desktop MCP setup",
           "instructions": "Open Settings > Developer > Edit Config, then add this section under the mcpServers key.",
           "fullExampleTitle": "Full config example",
@@ -1201,6 +1238,7 @@
         "exportFeed": {
           "title": "JSON export feed requests",
           "description": "Export feed URLs require an Authorization header with this token as a bearer token. Replace the feed URL placeholder with the Feed URL from the export page.",
+          "curlTitle": "cURL example",
           "pythonTitle": "Python example",
           "phpTitle": "PHP example"
         }
